Toolbox Fixes
Made the tool script possible to run in background that allows the tools to actually run, and not run out of memory.

Also, code now creates layers to clip instead of the the actual base layer.

Once this works, may try to revert backt to building the whole data from VRI (insect), and Fire base, as well as Hmn Distrubance.